(svn r1701) Support multi-feature callbacks where at least one of the features is 8 or higher.

+++ act123.h
@@ -136,7 +136,7 @@ private:
 	bool checks1C;
 };
 
-class Callbacks:public auto_array<uchar>{
+class Callbacks:public auto_array<uint>{
 public:
 	uint numcallbacks;
 	SINGLETON(Callbacks);
